+++ Session/Data/Entity/PlanetDomain/ResourceDeposit/ResourceDeposit.cs
@@ -9,16 +9,16 @@ public class ResourceDeposit : Entity
     public override EntityTypeTreeNode GetEntityTypeTreeNode() => EntityTypeTreeNode;
     public static EntityTypeTreeNode EntityTypeTreeNode { get; private set; }
     public ModelRef<Item> Item { get; protected set; }
-    public float Size { get; protected set; }
+    public int Size { get; protected set; }
     public EntityRef<MapPolygon> Poly { get; protected set; }
-    public static ResourceDeposit Create(Item resource, MapPolygon poly, float size, CreateWriteKey key)
+    public static ResourceDeposit Create(Item resource, MapPolygon poly, int size, CreateWriteKey key)
     {
         var d = new ResourceDeposit(key.IdDispenser.GetID(), resource.MakeRef(), poly.MakeRef(), size);
         key.Create(d);
         return d;
     }
 
-    [SerializationConstructor] private ResourceDeposit(int id, ModelRef<Item> item, EntityRef<MapPolygon> poly, float size) : base(id)
+    [SerializationConstructor] private ResourceDeposit(int id, ModelRef<Item> item, EntityRef<MapPolygon> poly, int size) : base(id)
     {
         Item = item;
         Size = size;
