+++ Session/Logic/Modules/Production/ProdAndConsumeProcedure.cs
@@ -64,8 +64,8 @@ public class ProdAndConsumeProcedure : Procedure
             {
                 var deposit = (ResourceDeposit) key.Data[kvp2.Key];
                 var loss = kvp2.Value;
-                var newSize = Mathf.Max(0f, deposit.Size - loss);
-                deposit.Set<float>(nameof(deposit.Size), newSize, key);
+                var newSize = Mathf.Max(0, deposit.Size - loss);
+                deposit.Set<int>(nameof(deposit.Size), newSize, key);
             }
         }
     }
